refactor(interfaces): consolidate type imports and remove deprecated types
- Merged type imports from '../types' into '../interfaces' for consistency across the codebase.
- Introduced a new 'chess-types.ts' file to define low-level chess domain literals, enhancing type organization.
- Removed the obsolete 'types/index.ts' file, streamlining the project structure and improving maintainability.
