+++ src/parsing/san-applier.ts
@@ -1,4 +1,4 @@
-import type { PieceToken } from '../types';
+import type { PieceToken } from '../interfaces';
 import { algebraicToCoordsAt } from '../board/board-coords';
 import type SanContext from './san-context';
