--- src/types/index.ts
@@ -1,5 +1,0 @@
-type PawnToken = 'a' | 'b' | 'c' | 'd' | 'e' | 'f' | 'g' | 'h';
-export type PieceToken = 'R' | 'N' | 'B' | 'Q' | 'K';
-
-export type Token = PieceToken | PawnToken | 'O';
-export type PlayerColor = 'b' | 'w';
